pyproject.toml
[project]
name = "urbansdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "databases>=0.9.0",
    "fastapi>=0.116.1",
    "geoalchemy2>=0.18.0",
    "geopandas>=1.1.1",
    "mapboxgl>=0.10.2",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pytest-asyncio>=1.1.0",
    "python-json-logger>=3.3.0",
    "requests>=2.32.4",
    "shapely>=2.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
[project.optional-dependencies]
test = ["pytest"]
dev = [
    "pytest",
    "ruff>=0.9.10",
    "black",
    "mypy",
    "trio",
    "pytest-asyncio",
    "pytest-mock",
    "httpx",
]
